development_requirements.txt
django==1.6.1
django-localflavor==1.0
south==0.8.4
django-grappelli==2.5.2
djangorestframework==2.3.12
django-extensions==1.3.3
Pillow==2.3.1
